Move disabled backend into its own package
