+++ traffic_ops/traffic_ops_golang/trafficvault/trafficvault.go
@@ -44,75 +44,6 @@ type TrafficVault interface {
 	Ping(tx *sql.Tx) (tc.TrafficVaultPingResponse, error)
 	GetBucketKey(bucket string, key string, tx *sql.Tx) ([]byte, bool, error)
 }
-type Error string
-
-func (e Error) Error() string {
-	return string(e)
-}
-
-const disabledErr = Error("traffic vault is disabled")
-
-type Disabled struct {
-}
-
-func (d *Disabled) GetDeliveryServiceSSLKeys(xmlID string, version string, tx *sql.Tx) (tc.DeliveryServiceSSLKeysV15, bool, error) {
-	return tc.DeliveryServiceSSLKeysV15{}, false, disabledErr
-}
-
-func (d *Disabled) PutDeliveryServiceSSLKeys(key tc.DeliveryServiceSSLKeys, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) DeleteDeliveryServiceSSLKeys(xmlID string, version string, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) DeleteOldDeliveryServiceSSLKeys(existingXMLIDs map[string]struct{}, cdnName string, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) GetCDNSSLKeys(cdnName string, tx *sql.Tx) ([]tc.CDNSSLKey, error) {
-	return nil, disabledErr
-}
-
-func (d *Disabled) GetDNSSECKeys(cdnName string, tx *sql.Tx) (tc.DNSSECKeysTrafficVault, bool, error) {
-	return nil, false, disabledErr
-}
-
-func (d *Disabled) PutDNSSECKeys(cdnName string, keys tc.DNSSECKeysTrafficVault, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) DeleteDNSSECKeys(cdnName string, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) GetURLSigKeys(xmlID string, tx *sql.Tx) (tc.URLSigKeys, bool, error) {
-	return nil, false, disabledErr
-}
-
-func (d *Disabled) PutURLSigKeys(xmlID string, keys tc.URLSigKeys, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) GetURISigningKeys(xmlID string, tx *sql.Tx) ([]byte, bool, error) {
-	return nil, false, disabledErr
-}
-
-func (d *Disabled) PutURISigningKeys(xmlID string, keysJson []byte, tx *sql.Tx) error {
-	return disabledErr
-}
-
-func (d *Disabled) DeleteURISigningKeys(xmlID string, tx *sql.Tx) error {
-	return disabledErr
-}
-func (d *Disabled) Ping(tx *sql.Tx) (tc.TrafficVaultPingResponse, error) {
-	return tc.TrafficVaultPingResponse{}, disabledErr
-}
-
-func (d *Disabled) GetBucketKey(bucket string, key string, tx *sql.Tx) ([]byte, bool, error) {
-	return nil, false, disabledErr
-}
 
 var backends = make(map[string]LoadFunc)
